requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
arize-phoenix-client==1.22.0
arize-phoenix-evals==2.5.0
arize-phoenix-otel==0.13.1
asyncer==0.0.10
attrs==25.4.0
audioop-lts==0.2.2
Authlib==1.6.5
backoff==2.2.1
bidict==0.23.1
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chainlit==2.8.4
charset-normalizer==3.4.4
chevron==0.14.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
faiss-cpu==1.12.0
fastapi==0.116.2
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
googleapis-common-protos==1.71.0
graphql-core==3.2.7
greenlet==3.2.4
grpc-interceptor==0.15.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.1
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-openai==1.0.2
langchain-pinecone==0.2.13
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
Lazify==0.4.0
literalai==0.1.201
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.20.0
mdurl==0.1.2
monotonic==1.6
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.4
openai==2.7.1
openinference-instrumentation==0.1.42
openinference-semantic-conventions==0.1.25
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-alephalpha==0.47.5
opentelemetry-instrumentation-anthropic==0.47.5
opentelemetry-instrumentation-bedrock==0.47.5
opentelemetry-instrumentation-chromadb==0.47.5
opentelemetry-instrumentation-cohere==0.47.5
opentelemetry-instrumentation-crewai==0.47.5
opentelemetry-instrumentation-google-generativeai==0.47.5
opentelemetry-instrumentation-groq==0.47.5
opentelemetry-instrumentation-haystack==0.47.5
opentelemetry-instrumentation-lancedb==0.47.5
opentelemetry-instrumentation-langchain==0.47.5
opentelemetry-instrumentation-llamaindex==0.47.5
opentelemetry-instrumentation-logging==0.59b0
opentelemetry-instrumentation-marqo==0.47.5
opentelemetry-instrumentation-mcp==0.47.5
opentelemetry-instrumentation-milvus==0.47.5
opentelemetry-instrumentation-mistralai==0.47.5
opentelemetry-instrumentation-ollama==0.47.5
opentelemetry-instrumentation-openai==0.47.5
opentelemetry-instrumentation-openai-agents==0.47.5
opentelemetry-instrumentation-pinecone==0.47.5
opentelemetry-instrumentation-qdrant==0.47.5
opentelemetry-instrumentation-redis==0.59b0
opentelemetry-instrumentation-replicate==0.47.5
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-sagemaker==0.47.5
opentelemetry-instrumentation-sqlalchemy==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-instrumentation-together==0.47.5
opentelemetry-instrumentation-transformers==0.47.5
opentelemetry-instrumentation-urllib3==0.59b0
opentelemetry-instrumentation-vertexai==0.47.5
opentelemetry-instrumentation-watsonx==0.47.5
opentelemetry-instrumentation-weaviate==0.47.5
opentelemetry-instrumentation-writer==0.47.5
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.3
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
ply==3.11
posthog==3.25.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.1.3
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.3
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.43.0
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlean.py==3.50.4.5
sse-starlette==3.0.3
starlette==0.48.0
strawberry-graphql==0.270.1
syncer==2.0.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tqdm==4.67.1
traceloop-sdk==0.47.5
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
